@@ -9,21 +9,21 @@
   ],
   "require": {
     "derhansen/sf_banners": "@dev",
-    "typo3/cms-core": "^12.4",
-    "php": ">=8.1"
+    "typo3/cms-core": "^13.4",
+    "php": ">=8.2"
   },
   "require-dev": {
     "typo3/cms-composer-installers": "^5.0",
-    "typo3/cms-backend": "^12.4",
-    "typo3/cms-frontend": "^12.4",
-    "typo3/cms-extbase": "^12.4",
-    "typo3/cms-felogin": "^12.4",
-    "typo3/cms-fluid-styled-content": "^12.4",
-    "typo3/cms-fluid": "^12.4",
-    "typo3/cms-tstemplate": "^12.4",
-    "typo3/cms-install": "^12.4",
-    "typo3/cms-extensionmanager": "^12.4",
-    "typo3/testing-framework": "^8.0",
+    "typo3/cms-backend": "^13.4",
+    "typo3/cms-frontend": "^13.4",
+    "typo3/cms-extbase": "^13.4",
+    "typo3/cms-felogin": "^13.4",
+    "typo3/cms-fluid-styled-content": "^13.4",
+    "typo3/cms-fluid": "^13.4",
+    "typo3/cms-tstemplate": "^13.4",
+    "typo3/cms-install": "^13.4",
+    "typo3/cms-extensionmanager": "^13.4",
+    "typo3/testing-framework": "^9.0",
     "friendsofphp/php-cs-fixer": "^3.12.0",
     "saschaegerer/phpstan-typo3": "^1.8.3",
     "phpstan/extension-installer": "^1.1"
